Added the new Export option 'Export ID Mapping Report' which is specifically designed for users who want to use ID Resolver to resolve the IDs, then modify each mod's config files manually. This exports everything in sections: each mod lists it's name, then it's blocks (sorted by original ID), then it's items (sorted by original ID).
Added support for FML (showing sprites, as well as checking for it's specific BaseMod type if it's there). The basemod thing is KIIIINDA experimental, but nobody responded whether it should be used or not, so I'm leaving it there right now.

Added support for showing terrain sprites remaining, not just the Item sprites.

Fixed one of the Labels not resizing correctly.

A few minor internal changes.